adding in serverboot v2 draft RFC.
* hurd/bootstrap.mdwn: I inlined the what_is_an_os_bootstrap page, and
wrote that the current bootstrap page is out of date and does not
include pci-arbiter or rumpdisk.
* hurd/what_is_an_os_bootstrap.mdwn: a new web page that is not meant
to be viewed directly.  Instead hurd/bootstrap and
open_issues/serverbootv2 is meant to inline the content.
* open_issues/serverbootv2.mdwn: Sergey proposed this new bootstrap
for the Hurd.  This is a draft RFC document that explains the
reasoning behind it.  Not that "Serverboot V2" is a working name.  We
have yet to find a better name for it.
Message-ID: <20240527010924.2607-1-jbranso@dismail.de>