sv.{gnu,nongnu}.org -> savannah.{gnu,nongnu}.org
This a very trivial change. The former is just an alias of latter.
But for new users and contributors like me, it sometimes confuses
them what the difference is between the two ones. Is one old and the
other new? Or something confusing else. So this patch helps them
save some time.

Signed-off-by: Yuqian Yang <crupest@crupest.life>
Message-ID: <8690a9c5-1da4-4adc-92b9-fd8f8dab2c88@crupest.life>